Add pragma.h header to centralize compiler warnings and fix sign conversion issues across codebase

+++ src/CommitDialog.cpp
@@ -162,7 +162,7 @@ void CommitDialog::on_pushButton_generate_with_ai_clicked()
 	diff_ = CommitMessageGenerator::make_diff(global->mainwindow->git(), commits_);
 	
 	auto [models, index] = global->appsettings.ai_models();
-	assert(index >= 0 && index < models.size());
+	assert(index >= 0 && (size_t)index < models.size());
 
 	GenerateCommitMessageDialog dlg(this, models, index);
 	dlg.setCommitIDs(commits_);
