+++ src/common/q/helper.h
@@ -14,9 +14,9 @@ public:
 	QS(QString const &s) : QString(s) {}
 	QS(QChar c) : QString(c) {}
 	QS(char const *p, int n = -1) : QString(QLatin1String(p, n)) {}
-	QS(std::string_view const &s) : QS(s.data(), s.size()) {}
+	QS(std::string_view const &s) : QS(s.data(), (int)s.size()) {}
 	QS(std::string const &s) : QString(QString::fromStdString(s)) {}
-	QS(std::vector<char> const &s) : QS(s.data(), s.size()) {}
+	QS(std::vector<char> const &s) : QS(s.data(), (int)s.size()) {}
 	operator std::string () const { return toStdString(); }
 	std::string ss() const { return toStdString(); }
 	QS &operator += (std::string const &s) { append(fromStdString(s)); return *this; }
@@ -28,9 +28,9 @@ public:
 	QBA(QByteArray const &ba) : QByteArray(ba) {}
 	QBA(char const *p, int n = -1) : QByteArray(p, n) {}
 	QBA(int n, char c) : QByteArray(n, c) {}
-	QBA(std::string const &s) : QBA(s.data(), s.size()) {}
-	QBA(std::string_view const &s) : QBA(s.data(), s.size()) {}
-	QBA(std::vector<char> const &s) : QBA(s.data(), s.size()) {}
+	QBA(std::string const &s) : QBA(s.data(), (int)s.size()) {}
+	QBA(std::string_view const &s) : QBA(s.data(), (int)s.size()) {}
+	QBA(std::vector<char> const &s) : QBA(s.data(), (int)s.size()) {}
 	std::vector<char> vc() const { return std::vector<char>(begin(), end()); }
 	std::string_view sv() const { return std::string_view(data(), size()); }
 	operator std::string_view () const { return sv(); }
